Move options to separate file
