+++ cmd/local/cmd.go
@@ -2,18 +2,11 @@ package cmd
 
 import (
 	"context"
-	"fmt"
-	"io/ioutil"
-	"os"
-	"strings"
 
-	"github.com/kyma-incubator/reconciler/internal/cli"
 	"github.com/kyma-incubator/reconciler/pkg/cluster"
-	file "github.com/kyma-incubator/reconciler/pkg/files"
 	"github.com/kyma-incubator/reconciler/pkg/keb"
 	"github.com/kyma-incubator/reconciler/pkg/logger"
 	"github.com/kyma-incubator/reconciler/pkg/reconciler"
-	"github.com/pkg/errors"
 
 	//Register all reconcilers
 	_ "github.com/kyma-incubator/reconciler/pkg/reconciler/instances"
@@ -26,78 +19,6 @@ import (
 const (
 	workspaceDir = ".workspace"
 )
-
-type Options struct {
-	*cli.Options
-	kubeconfigFile string
-	kubeconfig     string
-	components     []string
-}
-
-func defaultComponentList() []keb.Components {
-	defaultComponents := []string{"cluster-essentials", "istio-configuration@istio-system",
-		"certificates@istio-system", "logging", "tracing", "kiali", "monitoring", "eventing", "ory", "api-gateway",
-		"service-catalog", "service-catalog-addons", "rafter", "helm-broker", "cluster-users", "serverless",
-		"application-connector@kyma-integration"}
-	return componentsFromStrings(defaultComponents)
-
-}
-
-func NewOptions(o *cli.Options) *Options {
-	return &Options{o,
-		"",         // kubeconfigFile
-		"",         // kubeconfig
-		[]string{}, // components
-	}
-}
-func (o *Options) Kubeconfig() string {
-	return o.kubeconfig
-}
-
-func componentsFromStrings(list []string) []keb.Components {
-	var components []keb.Components
-	for _, item := range list {
-		s := strings.Split(item, "@")
-		name := s[0]
-		namespace := "kyma-system"
-		if len(s) >= 2 {
-			namespace = s[1]
-		}
-		components = append(components, keb.Components{Component: name, Namespace: namespace})
-	}
-	return components
-}
-
-func (o *Options) Components() []keb.Components {
-	components := componentsFromStrings(o.components)
-	if len(components) == 0 {
-		components = defaultComponentList()
-	}
-	return components
-}
-
-func (o *Options) Validate() error {
-	err := o.Options.Validate()
-	if err != nil {
-		return err
-	}
-	if o.kubeconfigFile == "" {
-		envKubeconfig, ok := os.LookupEnv("KUBECONFIG")
-		if !ok {
-			return fmt.Errorf("KUBECONFIG environment variable and kubeconfig flag is missing")
-		}
-		o.kubeconfigFile = envKubeconfig
-	}
-	if !file.Exists(o.kubeconfigFile) {
-		return fmt.Errorf("Reference kubeconfig file '%s' not found", o.kubeconfigFile)
-	}
-	content, err := ioutil.ReadFile(o.kubeconfigFile)
-	if err != nil {
-		return errors.Wrap(err, fmt.Sprintf("Failed to read kubeconfig file '%s'", o.kubeconfigFile))
-	}
-	o.kubeconfig = string(content)
-	return nil
-}
 
 func NewCmd(o *Options) *cobra.Command {
 	cmd := &cobra.Command{
@@ -113,6 +34,8 @@ func NewCmd(o *Options) *cobra.Command {
 	}
 	cmd.Flags().StringVar(&o.kubeconfigFile, "kubeconfig", "", "Path to kubeconfig file")
 	cmd.Flags().StringSliceVar(&o.components, "components", []string{}, "Comma separated list of components with optional namespace, e.g. serverless,certificates@istio-system,monitoring")
+	cmd.Flags().StringVar(&o.version, "version", "main", "Kyma version")
+	cmd.Flags().StringVar(&o.profile, "profile", "evaluation", "Kyma profile")
 	return cmd
 }
 
@@ -142,8 +65,8 @@ func RunLocal(o *Options) error {
 	ls, _ := scheduler.NewLocalScheduler(keb.Cluster{
 		Kubeconfig: o.kubeconfig,
 		KymaConfig: keb.KymaConfig{
-			Version:    "main",
-			Profile:    "evaluation",
+			Version:    o.version,
+			Profile:    o.profile,
 			Components: o.Components()}}, workerFactory, true)
 	return ls.Run(context.Background())
 }
